fix(sqlite, test): column overrides still have runtime type checking, use the correct Rust type to match the sql type

+++ tests/sqlite/macros.rs
@@ -100,7 +100,7 @@ async fn test_column_override_nullable() -> anyhow::Result<()> {
 
 #[derive(PartialEq, Eq, Debug, sqlx::Type)]
 #[sqlx(transparent)]
-struct MyInt(i32);
+struct MyInt(i64);
 
 #[sqlx_macros::test]
 async fn test_column_override_wildcard() -> anyhow::Result<()> {
